Add text notes to subfolders
These notes serve the following function: if someone clones into this repository, they will have the correct subfolder structure for the different functions to work. Some of them generate graphs and save them into a subfolder; others fit time-consuming models and save them as .Rda objects into another subfolder. I don't want the figures and R-data objects to be uploaded to the repo itself, but if the right path structure is not in place then many scripts will crash.